requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
rsa==4.9
Telethon==1.35.0
yarl==1.9.4
